// ts/features/itwallet/credentials/analytics/enum.ts
export enum ITW_CREDENTIALS_ERRORS_EVENTS {
  ITW_VAULT_CREDENTIAL_STORE_FAILED = "ITW_VAULT_CREDENTIAL_STORE_FAILED",
  ITW_VAULT_CREDENTIAL_REMOVE_FAILED = "ITW_VAULT_CREDENTIAL_REMOVE_FAILED",
  ITW_VAULT_MIGRATION_FAILED = "ITW_VAULT_MIGRATION_FAILED",
  ITW_VAULT_COHERENCE_CHECK_FAILED = "ITW_VAULT_COHERENCE_CHECK_FAILED",
  ITW_VAULT_ORPHANED_CREDENTIALS_FOUND = "ITW_VAULT_ORPHANED_CREDENTIALS_FOUND"
}

export enum ITW_CREDENTIALS_TECH_EVENTS {
  ITW_VAULT_MIGRATION_REQUEST = "ITW_VAULT_MIGRATION_REQUEST",
  ITW_VAULT_MIGRATION_SUCCESS = "ITW_VAULT_MIGRATION_SUCCESS"
}